+++ lib/views/expense_view.dart
@@ -7,6 +7,7 @@ import 'package:flutter/foundation.dart';
 import 'package:fl_chart/fl_chart.dart';
 import '../utils/money_utils.dart';
 import '../widgets/currency_text_field.dart';
+import '../widgets/keyboard_aware_padding.dart';
 import '../data/db_helper.dart';
 import '../services/notification_service.dart';
 import '../services/sync_service.dart';
@@ -595,12 +596,7 @@ class _ExpenseViewState extends State<ExpenseView> {
         builder: (ctx, setS) {
           titleC.addListener(() => setS(() {}));
           amountC.addListener(() => setS(() {}));
-          return Padding(
-            padding: EdgeInsets.only(
-              bottom:
-                  MediaQuery.viewInsetsOf(context).bottom +
-                  MediaQuery.paddingOf(context).bottom,
-            ),
+          return KeyboardAwarePadding(
             child: Container(
               decoration: const BoxDecoration(
                 color: PopupTheme.bgDark,
@@ -1315,12 +1311,7 @@ class _ExpenseViewState extends State<ExpenseView> {
         builder: (ctx, setS) {
           titleC.addListener(() => setS(() {}));
           amountC.addListener(() => setS(() {}));
-          return Padding(
-            padding: EdgeInsets.only(
-              bottom:
-                  MediaQuery.viewInsetsOf(context).bottom +
-                  MediaQuery.paddingOf(context).bottom,
-            ),
+          return KeyboardAwarePadding(
             child: Container(
               decoration: const BoxDecoration(
                 color: PopupTheme.bgDark,